requirements.txt
aiofile==3.8.8
aiohttp==3.9.1
aiolimiter==1.1.0
aiosignal==1.3.1
anyio==4.0.0
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.2.0
cachetools==5.3.2
caio==0.9.13
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.4
distro==1.9.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
jaraco.context==4.3.0
more-itertools==10.2.0
multidict==6.0.4
numpy==1.26.0
openpyxl==3.1.2
pandas==2.1.1
pdfminer.six==20221105
pdfplumber==0.10.2
Pillow==10.0.1
pycparser==2.21
PyPDF2==3.0.1
pypdfium2==4.20.0
python-dateutil==2.8.2
python-telegram-bot==20.8
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
tabula-py==2.9.0
tabulate==0.9.0
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
urllib3==2.1.0
wolframalpha-asyncio==4.0.1
xlrd==2.0.1
xmltodict==0.13.0
yarl==1.9.4
